+++ src/main/java/org/opentripplanner/routing/api/request/request/filter/TransitFilterRequest.java
@@ -2,50 +2,33 @@ package org.opentripplanner.routing.api.request.request.filter;
 
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import org.opentripplanner.framework.tostring.ToStringBuilder;
-import org.opentripplanner.transit.model.network.Route;
+import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.timetable.TripTimes;
 
 public class TransitFilterRequest implements Serializable, TransitFilter {
 
-  public static class Builder {
+  /**
+   * This is stored as an array, as they are iterated over for each trip when filtering transit
+   * data. Iterator creation is relatively expensive compared to iterating over a short array.
+   */
+  private final SelectRequest[] select;
 
-    private final List<SelectRequest> select = new ArrayList<>();
-    private final List<SelectRequest> not = new ArrayList<>();
-
-    public Builder addSelect(SelectRequest selectRequest) {
-      this.select.add(selectRequest);
-
-      return this;
-    }
-
-    public Builder addNot(SelectRequest selectRequest) {
-      this.not.add(selectRequest);
-
-      return this;
-    }
-
-    public TransitFilterRequest build() {
-      return new TransitFilterRequest(this);
-    }
-  }
+  /**
+   * {@link TransitFilterRequest#select}
+   */
+  private final SelectRequest[] not;
 
   public TransitFilterRequest(Builder builder) {
-    this.select = Collections.unmodifiableList(builder.select);
-    this.not = Collections.unmodifiableList(builder.not);
+    this.select = builder.select.toArray(SelectRequest[]::new);
+    this.not = builder.not.toArray(SelectRequest[]::new);
   }
-
-  private final List<SelectRequest> select;
-  private final List<SelectRequest> not;
 
   public List<SelectRequest> select() {
-    return select;
-  }
-
-  public List<SelectRequest> not() {
-    return not;
+    return Collections.unmodifiableList(Arrays.asList(select));
   }
 
   public static Builder of() {
@@ -75,15 +58,22 @@ public class TransitFilterRequest implements Serializable, TransitFilter {
   }
 
   @Override
-  public boolean matchRoute(Route route) {
-    if (!select.isEmpty()) {
-      if (select.stream().noneMatch(s -> s.matches(route))) {
+  public boolean matchTripPattern(TripPattern tripPattern) {
+    if (select.length != 0) {
+      var anyMatch = false;
+      for (SelectRequest s : select) {
+        if (s.matches(tripPattern)) {
+          anyMatch = true;
+          break;
+        }
+      }
+      if (!anyMatch) {
         return false;
       }
     }
 
-    if (!not.isEmpty()) {
-      if (not.stream().anyMatch(s -> s.matches(route))) {
+    for (SelectRequest s : not) {
+      if (s.matches(tripPattern)) {
         return false;
       }
     }
@@ -93,7 +83,7 @@ public class TransitFilterRequest implements Serializable, TransitFilter {
 
   @Override
   public boolean matchTripTimes(TripTimes tripTimes) {
-    if (!select.isEmpty()) {
+    if (select.length != 0) {
       var anyMatch = false;
       for (var selectRequest : select) {
         if (selectRequest.matches(tripTimes)) {
@@ -106,15 +96,8 @@ public class TransitFilterRequest implements Serializable, TransitFilter {
       }
     }
 
-    if (!not.isEmpty()) {
-      var anyMatch = false;
-      for (var selectRequest : select) {
-        if (selectRequest.matches(tripTimes)) {
-          anyMatch = true;
-          break;
-        }
-      }
-      if (anyMatch) {
+    for (SelectRequest s : not) {
+      if (s.matches(tripTimes)) {
         return false;
       }
     }
@@ -126,8 +109,30 @@ public class TransitFilterRequest implements Serializable, TransitFilter {
   public String toString() {
     return ToStringBuilder
       .of(TransitFilterRequest.class)
-      .addCol("select", select)
-      .addCol("not", not)
+      .addCol("select", Arrays.asList(select))
+      .addCol("not", Arrays.asList(not))
       .toString();
+  }
+
+  public static class Builder {
+
+    private final List<SelectRequest> select = new ArrayList<>();
+    private final List<SelectRequest> not = new ArrayList<>();
+
+    public Builder addSelect(SelectRequest selectRequest) {
+      this.select.add(selectRequest);
+
+      return this;
+    }
+
+    public Builder addNot(SelectRequest selectRequest) {
+      this.not.add(selectRequest);
+
+      return this;
+    }
+
+    public TransitFilterRequest build() {
+      return new TransitFilterRequest(this);
+    }
   }
 }