+++ src/test/java/org/opentripplanner/routing/algorithm/TestBanning.java
@@ -6,16 +6,14 @@ import static org.opentripplanner.transit.model._data.TransitModelForTest.id;
 
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.routing.algorithm.raptoradapter.transit.request.RouteRequestTransitDataProviderFilter;
-import org.opentripplanner.routing.api.request.request.TransitRequest;
 import org.opentripplanner.routing.api.request.request.filter.SelectRequest;
 import org.opentripplanner.routing.api.request.request.filter.TransitFilterRequest;
-import org.opentripplanner.routing.core.RouteMatcher;
 import org.opentripplanner.transit.model._data.TransitModelForTest;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.network.Route;
+import org.opentripplanner.transit.model.network.StopPattern;
+import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.organization.Agency;
 
 /**
@@ -26,55 +24,72 @@ public class TestBanning {
 
   @Test
   public void testSetBannedOnRequest() {
-    Collection<Route> routes = getTestRoutes();
+    Collection<TripPattern> patterns = getTestPatterns();
 
     var filterRequest = TransitFilterRequest
       .of()
-      .addNot(SelectRequest.of().withRoutes(RouteMatcher.parse("F__RUT:Route:1")).build())
-      .addNot(
-        SelectRequest.of().withAgencies(List.of(FeedScopedId.parseId("F:RUT:Agency:2"))).build()
-      )
+      .addNot(SelectRequest.of().withRoutes(List.of(id("RUT:Route:1"))).build())
+      .addNot(SelectRequest.of().withAgencies(List.of(id("RUT:Agency:2"))).build())
       .build();
 
-    Collection<FeedScopedId> bannedRoutes = RouteRequestTransitDataProviderFilter.bannedRoutes(
-      List.of(filterRequest),
-      routes
-    );
+    Collection<FeedScopedId> bannedPatterns = bannedPatterns(List.of(filterRequest), patterns);
 
-    assertEquals(2, bannedRoutes.size());
-    assertTrue(bannedRoutes.contains(id("RUT:Route:1")));
-    assertTrue(bannedRoutes.contains(id("RUT:Route:3")));
+    assertEquals(2, bannedPatterns.size());
+    assertTrue(bannedPatterns.contains(id("RUT:JourneyPattern:1")));
+    assertTrue(bannedPatterns.contains(id("RUT:JourneyPattern:3")));
   }
 
   @Test
   public void testSetWhiteListedOnRequest() {
-    Collection<Route> routes = getTestRoutes();
+    Collection<TripPattern> patterns = getTestPatterns();
 
     var filterRequest = TransitFilterRequest
       .of()
-      .addSelect(SelectRequest.of().withRoutes(RouteMatcher.parse("F__RUT:Route:1")).build())
-      .addSelect(
-        SelectRequest.of().withAgencies(List.of(FeedScopedId.parseId("F:RUT:Agency:2"))).build()
-      )
+      .addSelect(SelectRequest.of().withRoutes(List.of(id("RUT:Route:1"))).build())
+      .addSelect(SelectRequest.of().withAgencies(List.of(id("RUT:Agency:2"))).build())
       .build();
 
-    Collection<FeedScopedId> bannedRoutes = RouteRequestTransitDataProviderFilter.bannedRoutes(
-      List.of(filterRequest),
-      routes
-    );
+    Collection<FeedScopedId> bannedPatterns = bannedPatterns(List.of(filterRequest), patterns);
 
-    assertEquals(1, bannedRoutes.size());
-    assertTrue(bannedRoutes.contains(id("RUT:Route:2")));
+    assertEquals(1, bannedPatterns.size());
+    assertTrue(bannedPatterns.contains(id("RUT:JourneyPattern:2")));
   }
 
-  private List<Route> getTestRoutes() {
+  private List<TripPattern> getTestPatterns() {
     Agency agency1 = TransitModelForTest.agency("A").copy().withId(id("RUT:Agency:1")).build();
     Agency agency2 = TransitModelForTest.agency("B").copy().withId(id("RUT:Agency:2")).build();
 
+    Route route1 = TransitModelForTest.route("RUT:Route:1").withAgency(agency1).build();
+    Route route2 = TransitModelForTest.route("RUT:Route:2").withAgency(agency1).build();
+    Route route3 = TransitModelForTest.route("RUT:Route:3").withAgency(agency2).build();
+
+    final StopPattern stopPattern = TransitModelForTest.stopPattern(2);
     return List.of(
-      TransitModelForTest.route("RUT:Route:1").withAgency(agency1).build(),
-      TransitModelForTest.route("RUT:Route:2").withAgency(agency1).build(),
-      TransitModelForTest.route("RUT:Route:3").withAgency(agency2).build()
+      TransitModelForTest
+        .tripPattern("RUT:JourneyPattern:1", route1)
+        .withStopPattern(stopPattern)
+        .build(),
+      TransitModelForTest
+        .tripPattern("RUT:JourneyPattern:2", route2)
+        .withStopPattern(stopPattern)
+        .build(),
+      TransitModelForTest
+        .tripPattern("RUT:JourneyPattern:3", route3)
+        .withStopPattern(stopPattern)
+        .build()
     );
+  }
+
+  private static Collection<FeedScopedId> bannedPatterns(
+    List<TransitFilterRequest> filterRequest,
+    Collection<TripPattern> patterns
+  ) {
+    return patterns
+      .stream()
+      .filter(pattern ->
+        filterRequest.stream().noneMatch(filter -> filter.matchTripPattern(pattern))
+      )
+      .map(TripPattern::getId)
+      .toList();
   }
 }