+++ docs/Changelog.md
@@ -63,6 +63,17 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add ORCA Fares support with Fare by Leg [#4773](https://github.com/opentripplanner/OpenTripPlanner/pull/4773)
 - Fix TPoint generation for real-time generated patterns [#4787](https://github.com/opentripplanner/OpenTripPlanner/pull/4787)
 - Do not override network name in the vehicle rental service directory service [#4807](https://github.com/opentripplanner/OpenTripPlanner/pull/4807)
+- Improve resiliency of the Transmodel API [#4806](https://github.com/opentripplanner/OpenTripPlanner/pull/4806)
+- Update GBFS support to v2.3 [#4809](https://github.com/opentripplanner/OpenTripPlanner/pull/4809)
+- Use TripPatterns for filtering [#4820](https://github.com/opentripplanner/OpenTripPlanner/pull/4820)
+- Add Fares V2 distance-based computation [#4798](https://github.com/opentripplanner/OpenTripPlanner/pull/4798)
+- Remove all edges from stop vertex in island pruning [#4846](https://github.com/opentripplanner/OpenTripPlanner/pull/4846)
+- Filter functionality for GroupOfLines/GroupOfRoutes + TransmodelAPI m… [#4812](https://github.com/opentripplanner/OpenTripPlanner/pull/4812)
+- Mapping for maxAccessEgressDurationPerMode in Transmodel API [#4829](https://github.com/opentripplanner/OpenTripPlanner/pull/4829)
+- Use headsign from the original pattern in a realtime added pattern if the stop sequence is unchanged [#4845](https://github.com/opentripplanner/OpenTripPlanner/pull/4845)
+- Remove usage of apache commons [#4840](https://github.com/opentripplanner/OpenTripPlanner/pull/4840)
+- Remove RouteMatcher [#4821](https://github.com/opentripplanner/OpenTripPlanner/pull/4821)
+- Improve boarding location linking on platforms [#4852](https://github.com/opentripplanner/OpenTripPlanner/pull/4852)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)